scripts/requirements.txt
PyYAML>=6.0.1
openapi-spec-validator>=0.6.0
semver>=3.0.0
rich>=13.7.0